Make game playable on webpage with JS
Add displayControl module pattern to bridge HTML and JS, making
interactive and playable game on the webpage. The formControl module
pattern A checkRound function takes in name of the players and starts
the game checks the grid after every player move on whether the game is
over, either by a player winning or a tie game. A button appears once
the game is over to invite players to play again.